requirements.txt
redis==3.5.3
python-decouple==3.1